Moved out CurrentSenseValveMotorDirect
